+++ Orego/src/orego/response/WLSResponseList.java
@@ -60,7 +60,7 @@ public class WLSResponseList extends AbstractResponseList {
 		int start = random.nextInt(vacantPoints.size());
 		
 		int i = start;
-		double bestValue = PASS_WINS_PRIOR / PASS_RUNS_PRIOR;
+		double bestValue = (double) PASS_WINS_PRIOR / (double) PASS_RUNS_PRIOR;
 		
 		int bestMove = Coordinates.PASS;
 		do {
